Fix asynchronous particle addition bug
The simulation is meant to support adding and removing paricles dynamically.
There were twp issues with the scheme that existed before this commit.

Firstly, the MTLSharedEvent used to synchronize the CPU and the GPU did not
have its counter increase montonically; instead, it was left at a fixed value of
1 and 2 for EACH subsequent addition. In short, the value doesn't reset after each
commit of the command buffer it is signaled in.

Secondly, even after this was subsequently fixed, there was a GPU-side of the error.
The positions of the particles had changed, but the positions encoded onto the GPU had
not been subsequently updated in the asynchrnous test. Hence, there was division by 0 leading
to the flurry of NaNs observed.