Improve inverse square force calculation
Division by a runtime constant on a GPU is notoriously slow. This commit
modifies the force computing functions in the Metal source code taking into account
the slow nature of runtime division on the GPU.

The new process takes advantage of the (relatively) fast inverse square-root
computation function built in to the shading langauge. Instead of directly computing the
squared distance and normalizing the displacement vector joining the particles,
the function now computes the distance squared between the particles and performs
a single inverse square-root computation on the squared distance to give the inverse of the
distance between the particles. Then, taking advantage of ILP, multiplies this thrice
to get the inverse cube of the distance between the particles as desired