Enchance physics compute pipeline shader
The main physics kernel prior to this commit had several deficiencies. The functions the kernel
delegated to computed forces and not accelerations. This had a two-fold impact on performance. 1) The
GPU has to perform an additional multiplication on every force call when it should instead do a single multiplication
after calculating the acceleration on the particle, and 2) the force is ultimately divided by the particles mass
to recompute the acceleration, an operation notoriously slow with runtime values.