Implement threadgroup memory particle kernel
The initial implementation of the particle physics kernel using threadgroup memory
is included in this commit. A few other modifications were made to the way the program
accepts particles.

On the GPU, simdgroups are the "fundamental" unit. They are sets of threads executing in parallel
the same set of instructions. Accordingly, branching on GPUs can prove to be prohibitively expensive.
Requiring that the number of particles be a multiple of the size of a simdgroup (as of 2/19/21 is 64 on macOS)
greatly improves performance over compute kernels that branch when they would otherwise be using values outside
the domain. Since the specific counts are not very obvious according to the user, this restriction should not
prove substantial.

The Metal API calls are not yet implemented to run the kernel. A different buffer arrangement is needed in the
new scenario. Should the implementation prove to be more efficient than the "brute force" method, the latter will be discarded