Patch threadgroup memory particle kernel overwrite bug
The particle kernel `ThreadgroupParticleKernel`  erroneously did not ensure that working simdgroups in the last iteration through the array of the previous frame's particle data did not exceed the maximum allowed index, meaning there was an out-of-bounds read from device memory that was written to thread memory. This commit patches that error.

Further additions include the extraction of MTLCaptureManager from the MSParticleRenderer class, as the API requires that an entire command buffer (or multiple) be captured and not just the individual commands. Simd vectors are also opened to support iteration through their elements using the Sequence Swift protocol