Extract axis rendering and prepare shared memory ring buffers
Introduced in this commit is the abstraction of the rendering classes into the `MSRenderer` protocol, each of which provides a customized drawing function under the protocol method stub. More importantly, support has been added (though incompletely) to allow a ring buffer to share memory instead of creating multiple MTLBuffer instances. The alignment constraints are not working as expected with the constant address space, which has stricter alignment rules on macOS, but this will be patched in the next commit.

A bug in the particle renderer was also patched. The MSTileBuffer holding memory to the two particle buffers continued to switch read and write assignments while the simulation was paused, causing the particles to jiggle between the two different positions stored in the two buffers. Additionally, it was discovered that within the ThreadgroupParticleKernel conversions were made from ushort to uint implicitly. This commit makes that conversion explicit