Abstract rendering state and include indirect command buffers for rendering
This commit introduces an indirect command buffer to the particle rendering pipeline. The ICB is encoded on the CPU. Further changes will allow the compute pass portion of the particle rendering pipeline to be encoded concurrently on the GPU on iOS (this feature is unavailable on macOS).

The render states of each render object was brought into each class as a private struct for clarity and to improve encapsulation. Other improvements include removing the `capacity` parameter when encoding a single value into a buffer. Additionally, `MSMemory` now correct sets its length to that passed into the initializer (it didn't before).